refeactor
get rid of screensscripts.js and moved all code to related blocks
