+++ blocks/home-button/home-button.js
@@ -1,8 +1,6 @@
-import { loadScript } from '../../scripts/lib-franklin.js';
+import { loadScript } from "../../scripts/lib-franklin.js";
 export default function decorate(block) {
-    loadScript('/blocks/screenscript.js',{ type: 'text/javascript', charset: 'UTF-8' });
-    [...block.firstElementChild.children][0].textContent="Home";
+    [...block.firstElementChild.children][0].textContent=[...block.firstElementChild.children][0].children[0].textContent;
     [...block.firstElementChild.children][0].setAttribute("onclick","loadHome()");
-    block.style.display="none";
-
   }
+loadScript('/blocks/home-button/delayed.js',{ type: 'text/javascript', charset: 'UTF-8' });
