+++ blocks/kiosk-cards/kiosk-cards.js
@@ -1,4 +1,4 @@
-import { createOptimizedPicture } from '../../scripts/lib-franklin.js';
+import { createOptimizedPicture,loadScript } from '../../scripts/lib-franklin.js';
 
 export default function decorate(block) {
   /* change to ul, li */
@@ -27,6 +27,5 @@ export default function decorate(block) {
   ul.querySelectorAll('img').forEach((img) => img.closest('picture').replaceWith(createOptimizedPicture(img.src, img.alt, false, [{ width: '750' }])));
   block.textContent = '';
   block.append(ul);
-  
-
 }
+loadScript('/blocks/kiosk-cards/delayed.js',{ type: 'text/javascript', charset: 'UTF-8' });
