+++ blocks/place-order-button/delayed.js
@@ -1,0 +1,8 @@
+function showKioskCard()
+{
+    document.querySelector(".screens-sequence-carousel-container").style.display="none";
+    //document.querySelector(".place-order-button").style.display="none";
+    //document.querySelector(".home-button").style.display="block";
+    document.querySelector(".kiosk-cards-container").style.display="block";
+    //document.querySelector(".cart-button").style.display="block";
+}
