+++ blocks/screens-sequence-carousel/screens-sequence-carousel.js
@@ -1,3 +1,4 @@
+import { loadScript } from "../../scripts/lib-franklin.js";
 export default function decorate(block) {
   const buttons = document.createElement('div');
   buttons.className = 'carousel-buttons';
@@ -6,7 +7,6 @@ export default function decorate(block) {
     classes.forEach((e, j) => {
       row.children[j].classList.add(`screens-sequence-carousel-${e}`);
     });
-  });
-  
-  
+  });  
 }
+loadScript('/blocks/screens-sequence-carousel/delayed.js',{ type: 'text/javascript', charset: 'UTF-8' });
